refactor(adapters): split structured CSV adapter seams
Why:
- structured_csv had become a refactor-first hotspot and no longer matched the
  repo standard for small bounded adapter modules
- parity proof for schema validation, review generation, and normalization was
  too indirect when the adapter lived in one large file

What:
- replace the flat adapter module with package-local adapter, contracts,
  feedback, normalization, and validation modules under structured_csv/
- keep the public adapter manifest and discovery entry point stable while
  extracting row validation, review assembly, and balance/event construction
  into direct seams
- add direct adapter-local tests for header matching, required-field failures,
  outbound amount canonicalization, and invalid schema handling

Checks:
- uv run python -m tools.run_quality_gates --full-tests
