+++ src/crypto_reconciliation/adapters/sources/structured_csv/__init__.py
@@ -1,0 +1,5 @@
+"""Structured CSV source adapter package."""
+
+from .adapter import ADAPTER, StructuredCsvSourceAdapter
+
+__all__ = ("ADAPTER", "StructuredCsvSourceAdapter")
